internal/sqlfuncgen: new package
Move towards really generating files...
